sense_client/requirements.txt
pillow>=10.0
scikit-image>=0.22
numpy>=1.24
pytesseract>=0.3
requests>=2.31
